+++ src/Conditions/BaseCondition.ts
@@ -33,43 +33,87 @@ export abstract class BaseCondition<
   public abstract test(context: LabelerContext): Promise<boolean> | boolean;
 
   protected testStringValue(value: string, context: LabelerContext, predicate?: TPredicate): boolean {
+    let result: boolean = false;
+
     predicate = predicate || this.predicate;
 
     if (typeof predicate === 'string') {
-      return this.stringComparer(predicate, value);
+      result = this.stringComparer(predicate, value);
+    } else if (typeof predicate === 'function') {
+      result = predicate(value, context);
+    } else if (predicate instanceof RegExp) {
+      // When a regex has the global flag set, test() will advance the lastIndex of the regex.
+      // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/RegExp/test#Using_test()_on_a_regex_with_the_global_flag
+      predicate.lastIndex = 0;
+      result = predicate.test(value);
+    } else if (Array.isArray(predicate)) {
+      result = predicate.includes(value);
     }
 
-    if (typeof predicate === 'function') {
-      return predicate(value, context);
-    }
+    context.logger.debug(
+      '{class}.{method}: pr#{pullRequest.code}: the value "{value}" testing with predicate {type} yields the result "{result}"',
+      {
+        class: this.constructor.name,
+        method: this.test.name,
+        value,
+        type: typeof predicate,
+        ...(
+          predicate instanceof RegExp
+            ? {
+              regExp: predicate
+            }
+            : {}
+        ),
+        result,
+        pullRequest: context.pullRequest,
+      }
+    );
 
-    if (predicate instanceof RegExp) {
-      return predicate.test(value);
-    }
-
-    if (Array.isArray(predicate)) {
-      return predicate.includes(value);
-    }
-
-    throw new Error('The predicate type is not supported. Expectected a string, Array<string>, RegExp or Function.');
+    return result;
   }
 
   /**
    * Returns the specified result depending on the value of the nothing option.
    */
-  protected testResult(result: boolean): boolean {
-    const { nothing } = this.getOptions();
-
-    return (
-      result && !nothing
+  protected testResult(value: boolean, context: LabelerContext): boolean {
+    const { nothing } = this.getOptions(context);
+    const result = (
+      value && !nothing
     )
     || (
-      !result && !!nothing
+      !value && !!nothing
     );
+
+    context.logger.debug(
+      '{class}.{method}: pr#{pullRequest.code}: the value "{value}" when option nothing is "{nothing}" then result is "{result}""',
+      {
+        class: this.constructor.name,
+        method: this.test.name,
+        nothing,
+        value,
+        result,
+        pullRequest: context.pullRequest,
+      }
+    );
+
+    return result;
   }
 
-  protected getOptions(): ConditionOptionsValues<TConditionOptions> {
-    return (this.options?.['values'] || {}) as ConditionOptionsValues<TConditionOptions>;
+  protected getOptions(context: LabelerContext): ConditionOptionsValues<TConditionOptions> {
+    const result = (this.options?.['values'] || {}) as ConditionOptionsValues<TConditionOptions>;
+
+    context.logger.debug(
+      '{class}.{method}: pr#{pullRequest.code}: found {count} options',
+      {
+        class: this.constructor.name,
+        method: this.test.name,
+        options: result,
+        count: Object.keys(result).length,
+        pullRequest: context.pullRequest,
+      }
+    );
+
+    return result;
   }
 
 }