+++ src/Conditions/WhenCommitCountCondition.ts
@@ -16,32 +16,32 @@ export class WhenCommitCountCondition extends BaseCondition<DefaultPredicateType
       greaterThanOrEqualTo,
       lessThan,
       lessThanOrEqualTo,
-    } = super.getOptions();
+    } = super.getOptions(context);
 
     const { commits: count } = await context.pullRequest.statusInfo.get();
 
     // TODO: helper
     if (equal && count === equal) {
-      return this.testResult(true);
+      return this.testResult(true, context);
     }
 
     if (greaterThan && count > greaterThan) {
-      return this.testResult(true);
+      return this.testResult(true, context);
     }
 
     if (greaterThanOrEqualTo && count >= greaterThanOrEqualTo) {
-      return this.testResult(true);
+      return this.testResult(true, context);
     }
 
     if (lessThan && count < lessThan) {
-      return this.testResult(true);
+      return this.testResult(true, context);
     }
 
     if (lessThanOrEqualTo && count <= lessThanOrEqualTo) {
-      return this.testResult(true);
+      return this.testResult(true, context);
     }
 
-    return this.testResult(false);
+    return this.testResult(false, context);
   }
 
 }