+++ src/Interfaces/ILogger.ts
@@ -1,5 +1,16 @@
 export interface ILogger {
 
-  log(message?: any, ...optionalParams: Array<any>): void;
+  /**
+   * @deprecated Please use `info` instead.
+   */
+  log(message?: any, ...meta: Array<any>): void;
+
+  info(message: string, ...meta: Array<any>): void;
+
+  warning(message: string, ...meta: Array<any>): void;
+
+  error(message: string, ...meta: Array<any>): void;
+
+  debug(message: string, ...meta: Array<any>): void;
 
 }