+++ src/Types/index.ts
@@ -4,5 +4,4 @@ export { LabelerOptions } from './LabelerOptions';
 export { MergeDirection } from './MergeDirection';
 export { StringComparer, stringComparison } from './StringComparer';
 export { StringTestValue } from './StringTestValue';
-export { TaskContext } from './TaskContext';
 export { TaskFunction } from './TaskFunction';