+++ src/index.ts
@@ -1,13 +1,13 @@
-import { Auth, GitHubClient } from './ApiProviders';
+import { Auth, GitHubClient, IApiProviderClient } from './ApiProviders';
 import { IConfig } from './Interfaces';
 import { Labeler } from './Labeler';
 import { LabelerOptions } from './Types';
 
 export { IConfig } from './Interfaces';
 export { LabelerContext } from './LabelerContext';
-export { TaskContext } from './Types'; // TODO: Remove in future releases
+export { Auth, GitHubClient, IApiProviderClient };
 
-type Options = {
+type BaseRepositoryOptions = {
 
   /**
    * Pull Requests processing config.
@@ -17,43 +17,69 @@ type Options = {
   config: IConfig;
 
   /**
-   * Information for accessing the GitHub repository.
-   */
-  auth: Auth;
-
-  /**
    * Additional options.
    */
   options?: LabelerOptions;
 
 };
 
+type DefaultRepositoryOptions = BaseRepositoryOptions & {
+
+  /**
+   * Information for accessing the GitHub repository.
+   */
+  auth: Auth;
+
+  /**
+   * Allows to specify a custom API client instance.
+   */
+  client?: undefined;
+
+};
+
+type CustomRepositoryOptions = BaseRepositoryOptions & {
+
+  /**
+   * Information for accessing the GitHub repository.
+   */
+  auth?: undefined;
+
+  /**
+   * Allows to specify a custom API client instance.
+   */
+  client: IApiProviderClient;
+
+};
+
+export type RepositoryOptions = DefaultRepositoryOptions | CustomRepositoryOptions;
+
 export { createConfig } from './Config';
 
 /**
  * Applies the specified configuration to the specified repository.
  */
-export function fix(options: Options): Promise<void> {
+export function fix(options: RepositoryOptions): Promise<void> {
   return run(options, false);
 }
 
 /**
  * Tests the specified configuration without taking any action to modify the repository.
  */
-export function test(options: Options): Promise<void> {
+export function test(options: RepositoryOptions): Promise<void> {
   return run(options, true);
 }
 
-function run(options: Options, test: boolean): Promise<void> {
+function run(options: RepositoryOptions, test: boolean): Promise<void> {
   const {
     config,
     auth,
     options: labelerOptions,
+    client,
   } = options;
 
   const labeler = new Labeler(
     config,
-    new GitHubClient(auth),
+    client || new GitHubClient(auth),
     labelerOptions
   );