+++ tests/Conditions/WhenStateTest.ts
@@ -1,5 +1,6 @@
 import { expect } from 'chai';
 
+import { DefaultConditionOptions } from '../../src/ConditionOptions';
 import { WhenState } from '../../src/Conditions';
 import { LabelerContext } from '../../src/LabelerContext';
 import { firstPullRequest } from '../Resources';
@@ -17,6 +18,19 @@ describe('WhenState', () => {
     expect(when.test(context)).to.be.true;
   });
 
+  it('should return false for opened a pull request with the "nothing" options', (): void => {
+    const options = new DefaultConditionOptions(null);
+    const when = new WhenState('open', options);
+    const context = new LabelerContext({
+      pullRequest: firstPullRequest,
+      test: true,
+    });
+
+    options.nothing();
+
+    expect(when.test(context)).to.be.false;
+  });
+
   it('should return true for closed a pull request', (): void => {
     const when = new WhenState('closed');
     const context = new LabelerContext({
@@ -25,6 +39,19 @@ describe('WhenState', () => {
     });
 
     expect(when.test(context)).to.be.true;
+  });
+
+  it('should return false for closed a pull request with the "nothing" options', (): void => {
+    const options = new DefaultConditionOptions(null);
+    const when = new WhenState('closed', options);
+    const context = new LabelerContext({
+      pullRequest: closedPullRequest,
+      test: true,
+    });
+
+    options.nothing();
+
+    expect(when.test(context)).to.be.false;
   });
 
   it('should return false for merged a pull request', (): void => {
@@ -37,6 +64,19 @@ describe('WhenState', () => {
     expect(when.test(context)).to.be.false;
   });
 
+  it('should return true for merged a pull request with the "nothing" options', (): void => {
+    const options = new DefaultConditionOptions(null);
+    const when = new WhenState('closed', options);
+    const context = new LabelerContext({
+      pullRequest: mergedPullRequest,
+      test: true,
+    });
+
+    options.nothing();
+
+    expect(when.test(context)).to.be.true;
+  });
+
   it('should return true for merged a pull request', (): void => {
     const when = new WhenState('merged');
     const context = new LabelerContext({
@@ -45,6 +85,19 @@ describe('WhenState', () => {
     });
 
     expect(when.test(context)).to.be.true;
+  });
+
+  it('should return false for merged a pull request with the "nothing" options', (): void => {
+    const options = new DefaultConditionOptions(null);
+    const when = new WhenState('merged', options);
+    const context = new LabelerContext({
+      pullRequest: mergedPullRequest,
+      test: true,
+    });
+
+    options.nothing();
+
+    expect(when.test(context)).to.be.false;
   });
 
 });