+++ tests/UpdateValueTest.ts
@@ -2,12 +2,12 @@ import { expect } from 'chai';
 
 import { UpdateValue } from '../src/UpdateValue';
 
-describe('UpdateValueTest', () => {
+describe('UpdateValue', () => {
 
   it('the default counter value should be zero', (): void => {
     const updateValue = new UpdateValue<string>();
 
-    expect(0).equal(updateValue.counter);
+    expect(updateValue.counter).to.be.equal(0);
   });
 
   it('the counter value after setting the value should be equal to one', (): void => {
@@ -15,7 +15,7 @@ describe('UpdateValueTest', () => {
 
     updateValue.value = 'Hello World!';
 
-    expect(1).equal(updateValue.counter);
+    expect(updateValue.counter).to.be.equal(1);
   });
 
   it('the counter value after double setting the value should be equal to two', (): void => {
@@ -24,7 +24,7 @@ describe('UpdateValueTest', () => {
     updateValue.value = 'Hello World!';
     updateValue.value = 'Hold We Roll!';
 
-    expect(2).equal(updateValue.counter);
+    expect(updateValue.counter).to.be.equal(2);
   });
 
   it('the value should be "Hello World!"', (): void => {
@@ -32,7 +32,7 @@ describe('UpdateValueTest', () => {
 
     updateValue.value = 'Hello World!';
 
-    expect('Hello World!').equal(updateValue.value);
+    expect(updateValue.value).to.be.equal('Hello World!');
   });
 
 });